+++ packages/graph-schema-utils/src/model/index.ts
@@ -83,7 +83,8 @@ export class GraphSchema {
             new Property(
               property.token,
               PropertyType.fromJsonStruct(property.type),
-              property.mandatory
+              property.mandatory,
+              property.$id
             )
         )
       );
@@ -125,7 +126,8 @@ export class GraphSchema {
               new Property(
                 property.token,
                 PropertyType.fromJsonStruct(property.type),
-                property.mandatory
+                property.mandatory,
+                property.$id
               )
           )
         );
@@ -187,7 +189,7 @@ export class NodeObjectType {
   labels: NodeLabel[];
   properties: Property[];
 
-  constructor(id: string, labels: NodeLabel[], properties: Property[]) {
+  constructor(id: string, labels: NodeLabel[], properties: Property[] = []) {
     this.$id = id;
     this.labels = labels;
     this.properties = properties;
@@ -222,7 +224,7 @@ export class RelationshipObjectType {
     type: RelationshipType,
     from: NodeObjectType,
     to: NodeObjectType,
-    properties: Property[]
+    properties: Property[] = []
   ) {
     this.$id = id;
     this.type = type;
@@ -250,35 +252,45 @@ export class RelationshipObjectType {
   }
 }
 
+export type PropertyTypes = PropertyBaseType | PropertyArrayType;
 export class Property {
   token: string;
-  type: PropertyBaseType | PropertyArrayType;
-  mandatory: boolean;
+  type: PropertyTypes | PropertyTypes[];
+  mandatory: boolean | undefined;
+  $id: string | undefined;
 
   constructor(
     token: string,
     type: PropertyBaseType | PropertyArrayType,
-    mandatory?: boolean
+    mandatory?: boolean,
+    $id?: string
   ) {
     this.token = token;
     this.type = type;
     this.mandatory = mandatory || false;
+    this.$id = $id;
   }
   toJsonStruct() {
     const typeVal = Array.isArray(this.type)
       ? this.type.map((t) => t.toJsonStruct())
       : this.type.toJsonStruct();
-    return {
+    const out = {
       type: typeVal,
       token: this.token,
-      mandatory: this.mandatory,
     };
+    if (this.mandatory !== undefined) {
+      out["mandatory"] = this.mandatory;
+    }
+    if (this.$id !== undefined) {
+      out["$id"] = this.$id;
+    }
+    return out;
   }
 }
 
 export class PropertyBaseType {
-  type: PropertyTypes;
-  constructor(type: PropertyTypes) {
+  type: PrimitivePropertyTypes;
+  constructor(type: PrimitivePropertyTypes) {
     this.type = type;
   }
   toJsonStruct() {
@@ -316,7 +328,7 @@ export class PropertyType {
   }
 }
 
-type PropertyTypes =
+type PrimitivePropertyTypes =
   | "integer"
   | "string"
   | "float"