+++ packages/graph-schema-utils/test/validation/validation.test.js
@@ -31,5 +31,11 @@ describe("Validate if JSON-documen is a string", () => {
     );
     assert.throws(() => validateSchema(JSON_SCHEMA, NON_JSON), InputTypeError);
   });
-  
+  test("Handles optional id:s on properties", () => {
+    const schema = readFile(
+      path.resolve(__dirname, "./test-schemas/optional-id.json")
+    );
+    validateSchema(JSON_SCHEMA, schema);
+    assert.doesNotThrow(() => validateSchema(JSON_SCHEMA, schema));
+  });
 });